More warnings removed.

+++ src/Bpp/Seq/Container/SequenceContainerIterator.cpp
@@ -67,7 +67,7 @@ const Sequence* SimpleSequenceContainerIterator::nextSequence()
 
 bool SimpleSequenceContainerIterator::hasMoreSequences() const
 {
-	return currentPosition_ < static_cast<int>(sequences_->getNumberOfSequences());
+	return currentPosition_ < sequences_->getNumberOfSequences();
 }
 
 /******************************************************************************/
